ref(snapshot): move exporter out of `mod.rs`
